Update install scripts

+++ hosts/vanguard/scripts/run-install.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
 
-rm -rf /mnt/persist/nixos
+[ -d "/mnt/persist/nixos" ] && rm -rf /mnt/persist/nixos
+
 git clone https://github.com/Aquastias/Nixos-Dotfiles.git --depth=1 /mnt/persist/nixos
+
 nixos-install --root /mnt --flake /mnt/persist/nixos#vanguard
